webapp/requirements.txt
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
cffi==1.14.6
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security @ git+https://github.com/mattupstate/flask-security@674b18103fa8734aca71bbd084ea01e3709817ef
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
passlib==1.7.4
pkg_resources==0.0.0
psutil==5.9.0
pycparser==2.20
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.21
uWSGI==2.0.19.1
Werkzeug==2.0.1
WTForms==2.3.3
